+++ riverpie/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.21.0
 
 - feat: add `addon` to `ReduxAction`
+- feat: show execution time of `ReduxAction` in `RiverpieTracingPage`
 - **BREAKING**: change `ref.emitMessage` to `ref.message`
 
 ## 0.20.0